# schema/openapi.yaml
openapi: "3.0.3"
info:
  title: Slimeopolis API
  description: |
    Global Slime Emporium API — Cloudflare API Shield demonstration.
    Upload this schema to: Security → API Shield → Schema Validation.
  version: "1.0.0"
  contact:
    name: Slimeopolis API
    url: https://slimeopolis.example.com
servers:
  - url: https://slimeopolis.example.com/api
    description: Production (replace with your actual domain)

# ─── Security Schemes ───────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from /auth/login or /auth/register.
        API Shield Session Identifier: configure the Authorization header
        as the session identifier for Volumetric Abuse Detection.
  schemas:
...
    WholesaleOrderItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          example: prod_jp_001
        quantity:
          type: integer
          minimum: 10
          example: 50
    CreateWholesaleOrderRequest:
      type: object
      required: [company_name, tax_id, items]
      properties:
        company_name:
          type: string
          minLength: 2
          maxLength: 200
          example: Acme Slimes Ltd.
        tax_id:
          type: string
          minLength: 2
          maxLength: 50
          example: US-123456789
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/WholesaleOrderItem"
        notes:
          type: string
          maxLength: 1000
          example: Priority processing requested

# ─── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /wholesale/inventory:
    get:
      operationId: getWholesaleInventory
      summary: "[mTLS Required] View full inventory with stock levels"
      description: |
        API Shield: PRIMARY mTLS demonstration endpoint.
        Requires a valid Cloudflare-issued client certificate.
        Also: Volumetric Abuse Detection for inventory scraping.
        Schema Validation: query param types.
        Configure mTLS rule in: Security → API Shield → mTLS
        Host: your-domain.com, Path: /api/wholesale/*, Action: Block
      tags: [Wholesale]
      security:
        - BearerAuth: []
      parameters:
        - name: country
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
        - name: category
          in: query
          schema:
            type: string
            enum: [butter, cloud, clear, floam, crunchy, glossy, metallic, glitter, glow, clay]
        - name: min_stock
          in: query
          schema:
            type: integer
            minimum: 0
            default: 10
      responses:
        "200":
          description: Full inventory list with cert metadata
        "403":
          description: No valid client certificate or not a wholesale account
  /wholesale/bulk-order:
    post:
      operationId: createBulkOrder
      summary: "[mTLS Required] Place a wholesale bulk order"
      description: |
        API Shield: mTLS required + Schema Validation.
        Validates company_name, tax_id, items array (each with product_id
        and quantity ≥ 10 units). 30% wholesale discount applied.
      tags: [Wholesale]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWholesaleOrderRequest"
      responses:
        "201":
          description: Bulk order confirmed with 30% discount
        "403":
          description: No valid client certificate or not a wholesale account
        "422":
          description: Validation error (e.g. quantity < 10)
...